Basic code/tutorial
Basic code to control the brushless motors using simpleFOC.
Open-loop is without an encoder to check the position, so if we force the motor to turn, we can managed to do it.
Closed-loop is with the encoder, so we cannotforce to turn, and if we managed to do it, it will go back by itself to the desired position.